+++ packages/dart_edge_audio/rust/include/dart_edge_audio.h
@@ -7,6 +7,7 @@
 
 typedef struct NativeAudioBytesResult {
   NativeOwnedBytes bytes;
+  NativeOwnedBytes waveform;
   char* result_json;
 } NativeAudioBytesResult;